requirements.txt
flask
pydantic
web3
python-dotenv
requests
google-generativeai
slither-analyzer
pymysql

